requirements.txt
python-dotenv==1.0.1  # For reading environment variables stored in .env file
langchain==0.2.2
langchain-community==0.2.3
unstructured==0.14.4  # Document loading
chromadb==0.5.0  # Vector storage
openai==1.31.1  # For OpenAI API interaction (if needed for query generation)
tiktoken==0.7.0  # For OpenAI embeddings

# Install Hugging Face embeddings
sentence-transformers==2.2.2  # For HuggingFace embeddings (supports SentenceTransformers models)
huggingface_hub==0.10.1  # Ensure compatibility with sentence-transformers
